Changed price from int to double

+++ app/src/main/java/co/com/monkeymobile/product_searcher/data/source/remote/response/ItemBackendResponse.kt
@@ -12,7 +12,7 @@ data class ItemBackendResponse(
     val condition: String?,
     @SerializedName("thumbnail") val imageUrl: String?,
     @SerializedName("currency_id") val currency: String?,
-    val price: Int?
+    val price: Double?
 ) : Parcelable
 
 fun ItemBackendResponse.toItem() = if (
